terraform/functions/discover_existing/requirements.txt
functions-framework==3.*
google-cloud-container
pyyaml
google-cloud-logging
kubernetes
google-api-python-client
google-cloud-resource-manager
google-cloud-pubsub